requirements.txt
numpy==1.23.2
pandas
suffix-tree==0.0.7
graphviz==0.20.1
networkx==2.8.5
rich
biopython
snakemake